refactor: use nanostores for toasts store
