+++ src/components/Toast/index.ts
@@ -1,2 +1,2 @@
 export { ToastContainer } from './ToastContainer';
-export { addToast, updateToast, removeToast } from './toast';
+export { addToast, updateToast, removeToast } from './toastStore';
